refactor: relocate Project model and unify frontmatter parsing
- Relocated the Project class definition to lib/models/data_model.dart to resolve transitive dart:io import issues, making it client-safe.
- Centralized duplicated Jekyll-style YAML/Markdown parsing into a server-only helper lib/server/content_parser.dart.
- Refactored content.dart, projects_content.dart, and about.dart to use the new helper.
- Fixes #30
