+++ lib/models/data_model.dart
@@ -61,6 +61,42 @@ class Post {
   });
 }
 
+class Project {
+  final String id;
+  final String name;
+  final String repo; // e.g. kevmoo/pubviz
+  final String? pubPackage; // e.g. pubviz
+  final String? installCommand;
+  final bool featured;
+  final bool ignore;
+  final List<String> relatedPostPermalinks;
+  final String contentHtml;
+  final String? latestVersion;
+  final int? githubStars;
+  final String? pubUrl;
+  final String? githubUrl;
+  final String? lastReviewedSha;
+  final DateTime? lastReviewedAt;
+
+  Project({
+    required this.id,
+    required this.name,
+    required this.repo,
+    this.pubPackage,
+    this.installCommand,
+    this.featured = false,
+    this.ignore = false,
+    this.relatedPostPermalinks = const [],
+    required this.contentHtml,
+    this.latestVersion,
+    this.githubStars,
+    this.pubUrl,
+    this.githubUrl,
+    this.lastReviewedSha,
+    this.lastReviewedAt,
+  });
+}
+
 typedef SocialLink = ({String href, String title, String iconClass});
 
 const socialLinks = <SocialLink>[
